JobSearch/project/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
whitenoise==6.4.0
dj-database-url==0.5.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.0